+++ src/core/sdpsos/__init__.py
@@ -1,13 +1,8 @@
 from .sdpsos import SDPSOS, SOSProblem
 from .manifold import RootSubspace
 from .solution import SolutionSDP
-from ...sdp import (
-    SDPProblem,
-    SDPTransformation, SDPMatrixTransform, SDPRowMasking, SDPVectorTransform
-)
+from ...sdp import SDPProblem
 
 __all__ = [
-    'SDPSOS', 'SDPProblem', 'SOSProblem', 'RootSubspace',
-    'SDPTransformation', 'SDPMatrixTransform', 'SDPRowMasking', 'SDPVectorTransform',
-    'SolutionSDP'
+    'SDPSOS', 'SDPProblem', 'SOSProblem', 'RootSubspace', 'SolutionSDP'
 ]
